# lab-1/api/openapi.yaml
openapi: 3.1.0
info:
  title: api
  version: 1.0.0
  summary: The service under test for lab 1.
  description: |
    A deliberately badly-behaved HTTP service: it allocates memory and burns CPU
    on demand, so that cgroup limits and namespaces have something to act on.

    `/info` is the instrument for the isolation experiments — it reports what the
    process believes about itself (pid, uid, hostname, namespaced pids,
    capability sets, seccomp state, cgroup path), which is exactly what
    namespaces and dropped privileges change.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local instance (override the listen address with ADDR).
paths:
...
  /eat:
    get:
      operationId: eatMemory
      summary: Allocate N megabytes and hold them.
      description: |
        Allocates in 32 MB chunks and writes one byte per page, which is what
        faults the pages in and gets them charged to the cgroup. The memory is
        held until `/free` (or until the OOM killer takes the process).
      parameters:
        - name: mb
          in: query
          required: true
          description: Megabytes to allocate.
          schema:
            type: integer
            minimum: 1
          example: 256
      responses:
        "200":
          description: Allocated; reports the running total.
          content:
            text/plain:
              schema:
                type: string
              example: "held 256 MB\n"
        "400":
          description: Missing or non-positive `mb`.
          content:
            text/plain:
              schema:
                type: string
              example: "usage: /eat?mb=N  (N > 0)\n"
...
  /burn:
    get:
      operationId: burnCPU
      summary: Pin N cores in a busy loop.
      description: |
        Starts N burners, each locked to its own OS thread so the load is steady
        and CPU throttling counters stay readable. Returns immediately; the
        burners keep running until `/stop`.
      parameters:
        - name: n
          in: query
          required: false
          description: Number of cores to load.
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
      responses:
        "200":
          description: Burners started; reports the running total.
          content:
            text/plain:
              schema:
                type: string
              example: "burning on 1 core(s), GET /stop to stop\n"
        "400":
          description: Non-positive `n`.
          content:
            text/plain:
              schema:
                type: string
              example: "usage: /burn?n=N  (N > 0)\n"
...